Remove unused `assert` dependency (#14)
`assert` is a Node.js built-in module, so that will take priority over
the `assert` dependency whenever it is imported. `browserify` and other
bundlers will inject this package whenever it seems `assert` imported,
but they will use their own versions, not the one specified here. In
practice, this is entirely unused.