Update ESLint config (#22)
The ESLint config has been updated to v7.0.1. This includes the
addition of `prettier`, new lint scripts, and a variety of other rule
changes. Most changes were fixable using `yarn lint:fix`.

The only lint violation that required manual changes was the
"unexpected use of `beforeEach` for a single test" rule. Each use of
`beforeEach` for a single test was removed, with the relevant code
being moved directly into the test block.

This also required an adjustment to CI to ensure that the build occurs
before the linter is run. This is because the tests are linted, and
they import the `dist` directory.