Replace Ethers with @metamask/eth-query
It seems that a recent change to switch from `ethjs-query` to Ethers has
caused a regression: when the network is switched, `this.web3` still
points to the previous network. This happens because Web3Provider saves
a copy of the `sendAsync` method on the given provider, so even if
`nonce-tracker` is given a provider proxy, `this.web3` still has a
reference to the `sendAsync` method from the previous provider object.

This is a common problem with Ethers. Although Ethers has some nice
features, we don't need to pull it in here, and we can use
`@metamask/eth-query` (which provides TypeScript support, unlike
`@metamask/ethjs-query`) instead. We do lose one feature of Ethers,
which is that it automatically converts response data for requests to
the appropriate types, but we can make use of `@metamask/utils` to help
with this.